+++ crates/tauri-build/src/lib.rs
@@ -263,7 +263,7 @@ impl WindowsAttributes {
     }
   }
 
-  /// Creates the default attriute set wihtou the default app manifest.
+  /// Creates the default attribute set without the default app manifest.
   #[must_use]
   pub fn new_without_app_manifest() -> Self {
     Self {
@@ -499,7 +499,7 @@ pub fn try_build(attributes: Attributes) -> Result<()> {
   println!("cargo:rustc-env=TAURI_ANDROID_PACKAGE_NAME_PREFIX={android_package_prefix}");
 
   if let Some(project_dir) = env::var_os("TAURI_ANDROID_PROJECT_PATH").map(PathBuf::from) {
-    mobile::generate_gradle_files(project_dir, &config)?;
+    mobile::generate_gradle_files(project_dir)?;
   }
 
   cfg_alias("dev", is_dev());