+++ crates/tauri-cli/src/mobile/ios/mod.rs
@@ -49,6 +49,7 @@ use std::{
 mod build;
 mod dev;
 pub(crate) mod project;
+mod run;
 mod xcode_script;
 
 pub const APPLE_DEVELOPMENT_TEAM_ENV_VAR_NAME: &str = "APPLE_DEVELOPMENT_TEAM";
@@ -95,6 +96,7 @@ enum Commands {
   Init(InitOptions),
   Dev(dev::Options),
   Build(build::Options),
+  Run(run::Options),
   #[clap(hide(true))]
   XcodeScript(xcode_script::Options),
 }
@@ -113,7 +115,8 @@ pub fn command(cli: Cli, verbosity: u8) -> Result<()> {
       )?
     }
     Commands::Dev(options) => dev::command(options, noise_level)?,
-    Commands::Build(options) => build::command(options, noise_level)?,
+    Commands::Build(options) => build::command(options, noise_level).map(|_| ())?,
+    Commands::Run(options) => run::command(options, noise_level)?,
     Commands::XcodeScript(options) => xcode_script::command(options)?,
   }
 
@@ -486,42 +489,6 @@ fn inject_resources(config: &AppleConfig, tauri_config: &TauriConfig) -> Result<
   }
 
   Ok(())
-}
-
-enum PlistKind {
-  Path(PathBuf),
-  Plist(plist::Value),
-}
-
-impl From<PathBuf> for PlistKind {
-  fn from(p: PathBuf) -> Self {
-    Self::Path(p)
-  }
-}
-impl From<plist::Value> for PlistKind {
-  fn from(p: plist::Value) -> Self {
-    Self::Plist(p)
-  }
-}
-
-fn merge_plist(src: Vec<PlistKind>) -> Result<plist::Value> {
-  let mut merged_plist = plist::Dictionary::new();
-
-  for plist_kind in src {
-    let plist = match plist_kind {
-      PlistKind::Path(p) => plist::Value::from_file(p).context("failed to read plist file"),
-      PlistKind::Plist(v) => Ok(v),
-    };
-    if let Ok(src_plist) = plist {
-      if let Some(dict) = src_plist.into_dictionary() {
-        for (key, value) in dict {
-          merged_plist.insert(key, value);
-        }
-      }
-    }
-  }
-
-  Ok(plist::Value::Dictionary(merged_plist))
 }
 
 pub fn signing_from_env() -> Result<(